% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmcmenet-package.R
\docType{package}
\name{glmcmenet-package}
\alias{glmcmenet-package}
\title{glmcmenet: Conditional Main-Effect Regularization for GLMs}
\description{
\code{glmcmenet} performs variable selection of conditional main effects (CMEs)
via a adaptive bi-level penalization framework for generalized linear models.
}
\section{C++ integration}{

The package uses Rcpp/RcppArmadillo. Native routines are registered.
}

\references{
Xie, K., & Xinwei, Deng. (2025). Adaptive Bi-Level Variable Selection of Conditional
Main Effects for Generalized Linear Models. \emph{Technometrics}, to be appear.
Mak, S., & Wu, C. J. (2019). cmenet: A new method for bi-level variable selection
of conditional main effects. \emph{Journal of the American Statistical Association}, 114(526), 844-856.
}
\keyword{internal}
